@@ -18,25 +18,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: "2.12.0-259.9.beta"
 
 dependencies:
   flutter:
     sdk: flutter
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.1
   toggle_switch: ^0.1.8
   http: ^0.12.0+4
   uuid: 2.2.2
   intl: ^0.16.1
-  google_maps_flutter: ^0.5.30
-  google_map_location_picker: ^3.3.4
-  flutter_localizations:
-    sdk: flutter
-
 
 dev_dependencies:
   flutter_test:
